move helpers.DirUpated to cmp.DirUpdated
