+++ helpers/helpers_test.go
@@ -294,106 +294,3 @@ func compareDirs(c *C, srcDir, destDir string) {
 	c.Assert(err, IsNil)
 	c.Assert(string(c1), Equals, string(c2))
 }
-
-func (ts *HTestSuite) TestSyncDirs(c *C) {
-
-	for _, l := range [][2]string{
-		[2]string{"src-short", "dst-loooooooooooong"},
-		[2]string{"src-loooooooooooong", "dst-short"},
-		[2]string{"src-eq", "dst-eq"},
-	} {
-
-		// ensure we have src, dest dirs with different length
-		srcDir := filepath.Join(c.MkDir(), l[0])
-		err := os.MkdirAll(srcDir, 0755)
-		c.Assert(err, IsNil)
-		destDir := filepath.Join(c.MkDir(), l[1])
-		err = os.MkdirAll(destDir, 0755)
-		c.Assert(err, IsNil)
-
-		// add a src subdir
-		subdir := filepath.Join(srcDir, "subdir")
-		err = os.Mkdir(subdir, 0755)
-		c.Assert(err, IsNil)
-		makeTestFiles(c, subdir, destDir)
-
-		// add a dst subdir that needs to get deleted
-		subdir2 := filepath.Join(destDir, "to-be-deleted-subdir")
-		err = os.Mkdir(subdir2, 0755)
-		subdir3 := filepath.Join(subdir2, "to-be-deleted-sub-subdir")
-		err = os.Mkdir(subdir3, 0755)
-
-		// and a toplevel
-		makeTestFiles(c, srcDir, destDir)
-
-		// do it
-		err = RSyncWithDelete(srcDir, destDir)
-		c.Assert(err, IsNil)
-
-		// ensure meta-data is identical
-		compareDirs(c, srcDir, destDir)
-		compareDirs(c, filepath.Join(srcDir, "subdir"), filepath.Join(destDir, "subdir"))
-	}
-}
-
-func (ts *HTestSuite) TestSyncDirFails(c *C) {
-	srcDir := c.MkDir()
-	err := os.MkdirAll(srcDir, 0755)
-	c.Assert(err, IsNil)
-
-	destDir := c.MkDir()
-	err = os.MkdirAll(destDir, 0755)
-	c.Assert(err, IsNil)
-
-	err = ioutil.WriteFile(filepath.Join(destDir, "meep"), []byte(nil), 0644)
-	c.Assert(err, IsNil)
-
-	// ensure remove fails
-	err = os.Chmod(destDir, 0100)
-	c.Assert(err, IsNil)
-	// make tempdir cleanup work again
-	defer os.Chmod(destDir, 0755)
-
-	// do it
-	err = RSyncWithDelete(srcDir, destDir)
-	c.Check(err, NotNil)
-	c.Check(err, ErrorMatches, ".*permission denied.*")
-}
-
-func (ts *HTestSuite) TestCopyIfDifferent(c *C) {
-	srcDir := c.MkDir()
-	dstDir := c.MkDir()
-
-	// new file
-	src := filepath.Join(srcDir, "bop")
-	dst := filepath.Join(dstDir, "bob")
-	err := ioutil.WriteFile(src, []byte(nil), 0644)
-	c.Assert(err, IsNil)
-
-	err = CopyIfDifferent(src, dst)
-	c.Assert(err, IsNil)
-	c.Check(FilesAreEqual(dst, src), Equals, true)
-
-	// updated file
-	src = filepath.Join(srcDir, "bip")
-	dst = filepath.Join(dstDir, "bib")
-	err = ioutil.WriteFile(src, []byte("123"), 0644)
-	c.Assert(err, IsNil)
-	err = ioutil.WriteFile(dst, []byte(nil), 0644)
-	c.Assert(err, IsNil)
-
-	err = CopyIfDifferent(src, dst)
-	c.Assert(err, IsNil)
-	c.Check(FilesAreEqual(dst, src), Equals, true)
-}
-
-func (ts *HTestSuite) TestCopyIfDifferentErrorsOnNoSrc(c *C) {
-	srcDir := c.MkDir()
-	dstDir := c.MkDir()
-
-	src := filepath.Join(srcDir, "mop")
-	dst := filepath.Join(dstDir, "mop")
-
-	err := CopyIfDifferent(src, dst)
-	c.Assert(err, NotNil)
-}
